+++ tests/Functional/BaseTestCase.php
@@ -16,6 +16,7 @@ use Sulu\Component\DocumentManager\Tests\Bootstrap;
 abstract class BaseTestCase extends \PHPUnit_Framework_TestCase
 {
     const BASE_NAME = 'test';
+
     const BASE_PATH = '/test';
 
     private $container;