+++ tests/Functional/Model/Mapping5Document.php
@@ -14,8 +14,12 @@ namespace Sulu\Component\DocumentManager\Tests\Functional\Model;
 class Mapping5Document
 {
     public $one;
+
     public $two;
+
     public $three;
+
     public $four;
+
     public $five;
 }